refactor: remove database/schemas
