+++ apps/electron/src/main/infrastructure/database/repositories/session/session-repository.ts
@@ -2,12 +2,36 @@ import { BaseRepository } from "../../core/base-repository";
 import { SqliteManager } from "../../core/sqlite-manager";
 import { v4 as uuidv4 } from "uuid";
 import { LocalChatSession, LocalSessionStatus } from "@mcp_router/shared";
-import { CHAT_SESSIONS_SCHEMA } from "../../schema/tables/chat-sessions";
 
 /**
  * Chat session repository for local database storage
  */
 export class SessionRepository extends BaseRepository<LocalChatSession> {
+  /**
+   * テーブル作成SQL
+   */
+  private static readonly CREATE_TABLE_SQL = `
+    CREATE TABLE IF NOT EXISTS chat_sessions (
+      id TEXT PRIMARY KEY,
+      agent_id TEXT NOT NULL,
+      title TEXT,
+      messages TEXT NOT NULL DEFAULT '[]',
+      created_at TEXT NOT NULL DEFAULT (datetime('now')),
+      updated_at TEXT NOT NULL DEFAULT (datetime('now')),
+      status TEXT NOT NULL DEFAULT 'pending' CHECK(status IN ('pending', 'processing', 'completed', 'failed')),
+      source TEXT NOT NULL DEFAULT 'ui'
+    )
+  `;
+
+  /**
+   * インデックス作成SQL
+   */
+  private static readonly INDEXES = [
+    "CREATE INDEX IF NOT EXISTS idx_chat_sessions_agent_id ON chat_sessions(agent_id)",
+    "CREATE INDEX IF NOT EXISTS idx_chat_sessions_created ON chat_sessions(created_at)",
+    "CREATE INDEX IF NOT EXISTS idx_chat_sessions_status ON chat_sessions(status)"
+  ];
+
   constructor(db: SqliteManager) {
     super(db, "chat_sessions");
     console.log(
@@ -21,15 +45,13 @@ export class SessionRepository extends BaseRepository<LocalChatSession> {
    */
   protected initializeTable(): void {
     try {
-      // スキーマ定義を使用してテーブルを作成
-      this.db.execute(CHAT_SESSIONS_SCHEMA.createSQL);
+      // テーブルを作成
+      this.db.execute(SessionRepository.CREATE_TABLE_SQL);
 
-      // スキーマ定義からインデックスを作成
-      if (CHAT_SESSIONS_SCHEMA.indexes) {
-        CHAT_SESSIONS_SCHEMA.indexes.forEach((indexSQL) => {
-          this.db.execute(indexSQL);
-        });
-      }
+      // インデックスを作成
+      SessionRepository.INDEXES.forEach((indexSQL) => {
+        this.db.execute(indexSQL);
+      });
 
       console.log("[SessionRepository] テーブルの初期化が完了しました");
     } catch (error) {
