--- apps/electron/src/main/infrastructure/database/schema/tables/index.ts
@@ -1,12 +1,0 @@
-/**
- * スキーマ定義のエクスポート
- */
-export * from "./agents";
-export * from "./deployed-agents";
-export * from "./servers";
-export * from "./request-logs";
-export * from "./chat-sessions";
-export * from "./settings";
-export * from "./migrations";
-export * from "./workspaces";
-export * from "./tokens";
