--- apps/electron/src/main/infrastructure/database/schema/tables/migrations.ts
@@ -1,14 +1,0 @@
-import { DatabaseTableSchema } from "@mcp_router/shared";
-
-/**
- * マイグレーションテーブルのスキーマ定義
- */
-export const MIGRATIONS_SCHEMA: DatabaseTableSchema = {
-  createSQL: `
-    CREATE TABLE IF NOT EXISTS migrations (
-      id TEXT PRIMARY KEY,
-      executed_at INTEGER NOT NULL
-    )
-  `,
-  indexes: [],
-};
